+++ Sources/NablaScheduling/Domain/Interactors/FetchConsentsAppointment/FetchConsentsInteractor.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-protocol FetchConcentsInteractor {
-    /// - Throws: ``NablaError``
-    func execute(location: LocationType) async throws -> Consents
-}