+++ Sources/NablaScheduling/Domain/Interactors/WatchConsentsAppointment/WatchConsentsInteractor.swift
@@ -1,0 +1,7 @@
+import Combine
+import Foundation
+import NablaCore
+
+protocol WatchConsentsInteractor {
+    func execute(location: LocationType) -> AnyPublisher<Consents, NablaError>
+}